Move BlockDigest struct to Libplanet.Store
